confirm record being sucessfully deleted
When a record is destroyed with `destroyRecord()` and a following
`save()`, it enters a in-flight state.

The adapter (we in this case) must notify the record that its deletion
was successful so we can `unloadRecord()` it and it may  `free()` itself
in the end.

If we don't do this, we get the following error:

  Assertion Failed: You can only unload a record which is not inFlight.